Move new application entity to maat module
